+++ src/com/company/TotalWithTax.java
@@ -8,7 +8,7 @@ Scanner tempo=new Scanner(System.in);
         System.out.println("Enter the price of your item");
 
         final double TAXRATE = 0.0825;
-        double price = tempo.nextInt();
+        double price = tempo.nextDouble();
         double salestax = price*TAXRATE;
         double total = price+salestax;
         System.out.println("Price: $"+price);
